Alterações de consulta no conselho de classe
