+++ src/SME.SGP.Aplicacao/Consultas/ConsultasConselhoClasse.cs
@@ -12,7 +12,7 @@ namespace SME.SGP.Aplicacao
 {
     public class ConsultasConselhoClasse : IConsultasConselhoClasse
     {
-        private readonly IRepositorioConselhoClasse repositorioConselhoClasse;
+        private readonly IRepositorioConselhoClasseConsulta repositorioConselhoClasseConsulta;
         private readonly IRepositorioPeriodoEscolar repositorioPeriodoEscolar;
         private readonly IRepositorioParametrosSistema repositorioParametrosSistema;
         private readonly IRepositorioConselhoClasseAluno repositorioConselhoClasseAluno;
@@ -25,7 +25,7 @@ namespace SME.SGP.Aplicacao
         private readonly IRepositorioTipoCalendario repositorioTipoCalendario;
         private readonly IMediator mediator;
 
-        public ConsultasConselhoClasse(IRepositorioConselhoClasse repositorioConselhoClasse,
+        public ConsultasConselhoClasse(IRepositorioConselhoClasseConsulta repositorioConselhoClasse,
                                        IRepositorioPeriodoEscolar repositorioPeriodoEscolar,
                                        IRepositorioParametrosSistema repositorioParametrosSistema,
                                        IRepositorioConselhoClasseAluno repositorioConselhoClasseAluno,
@@ -38,7 +38,7 @@ namespace SME.SGP.Aplicacao
                                        IServicoDeNotasConceitos servicoDeNotasConceitos,
                                        IMediator mediator)
         {
-            this.repositorioConselhoClasse = repositorioConselhoClasse ?? throw new ArgumentNullException(nameof(repositorioConselhoClasse));
+            this.repositorioConselhoClasseConsulta = repositorioConselhoClasse ?? throw new ArgumentNullException(nameof(repositorioConselhoClasse));
             this.repositorioPeriodoEscolar = repositorioPeriodoEscolar ?? throw new ArgumentNullException(nameof(repositorioPeriodoEscolar));
             this.repositorioParametrosSistema = repositorioParametrosSistema ?? throw new ArgumentNullException(nameof(repositorioParametrosSistema));
             this.repositorioConselhoClasseAluno = repositorioConselhoClasseAluno ?? throw new ArgumentNullException(nameof(repositorioConselhoClasseAluno));
@@ -83,7 +83,7 @@ namespace SME.SGP.Aplicacao
             if (fechamentoTurma == null && !turma.EhAnoAnterior())
                 throw new NegocioException("Fechamento da turma não localizado " + (!ehFinal && bimestre > 0 ? $"para o bimestre {bimestre}" : ""));
 
-            var conselhoClasse = fechamentoTurma != null ? await repositorioConselhoClasse.ObterPorFechamentoId(fechamentoTurma.Id) : null;
+            var conselhoClasse = fechamentoTurma != null ? await repositorioConselhoClasseConsulta.ObterPorFechamentoId(fechamentoTurma.Id) : null;
 
             var periodoEscolarId = fechamentoTurma?.PeriodoEscolarId;
 
@@ -169,7 +169,7 @@ namespace SME.SGP.Aplicacao
         }
 
         public ConselhoClasse ObterPorId(long conselhoClasseId)
-            => repositorioConselhoClasse.ObterPorId(conselhoClasseId);
+            => repositorioConselhoClasseConsulta.ObterPorId(conselhoClasseId);
 
         public async Task<(int, bool)> ValidaConselhoClasseUltimoBimestre(Turma turma)
         {
