+++ src/SME.SGP.Aplicacao/Queries/DashboardFechamento/ObterConselhoClasseSituacao/ObterConselhoClasseSituacaoQueryHandler.cs
@@ -10,15 +10,15 @@ namespace SME.SGP.Aplicacao
 {
     public class ObterConselhoClasseSituacaoQueryHandler : IRequestHandler<ObterConselhoClasseSituacaoQuery, IEnumerable<ConselhoClasseSituacaoQuantidadeDto>>
     {
-        private readonly IRepositorioConselhoClasse repositorio;
+        private readonly IRepositorioConselhoClasseConsulta repositorioConselhoClasseConsulta;
 
         public ObterConselhoClasseSituacaoQueryHandler(IRepositorioConselhoClasse repositorio)
         {
-            this.repositorio = repositorio ?? throw new ArgumentNullException(nameof(repositorio));
+            this.repositorioConselhoClasseConsulta = repositorioConselhoClasseConsulta ?? throw new ArgumentNullException(nameof(repositorioConselhoClasseConsulta));
         }
         public async Task<IEnumerable<ConselhoClasseSituacaoQuantidadeDto>> Handle(ObterConselhoClasseSituacaoQuery request, CancellationToken cancellationToken)
         {
-           return await repositorio.ObterConselhoClasseSituacao(request.UeId,
+           return await repositorioConselhoClasseConsulta.ObterConselhoClasseSituacao(request.UeId,
                                                                 request.Ano, 
                                                                 request.DreId,
                                                                 request.Modalidade,
