+++ src/SME.SGP.Dados/Repositorios/RepositorioTurmaConsulta.cs
@@ -19,5 +19,16 @@ namespace SME.SGP.Dados.Repositorios
             return await contexto.Conexao.QueryFirstOrDefaultAsync<Turma>("select * from turma where turma_id = @turmaCodigo", new { turmaCodigo });
         }
 
+        public async Task<string> ObterTurmaCodigoPorConselhoClasseId(long conselhoClasseId)
+        {
+            var query = @"select t.turma_id
+                          from conselho_classe cc
+                          inner join fechamento_turma ft on ft.id = cc.fechamento_turma_id
+                          inner join turma t on t.id = ft.turma_id
+                         where not cc.excluido and not ft.excluido
+                           and cc.id = @conselhoClasseId";
+
+            return await contexto.Conexao.QueryFirstOrDefaultAsync<string>(query, new { conselhoClasseId });
+        }
     }
 }
