+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioTurmaConsulta.cs
@@ -5,5 +5,6 @@ namespace SME.SGP.Dominio.Interfaces
     public interface IRepositorioTurmaConsulta
     {
         Task<Turma> ObterPorCodigo(string turmaCodigo);
+        Task<string> ObterTurmaCodigoPorConselhoClasseId(long conselhoClasseId);
     }
 }
